Finished changing imports and solving err
