+++ .gitignore
@@ -3,3 +3,4 @@
 .env.dev
 config.json
 node_modules
+.github
