+++ src/handlers/eventHandler.ts
@@ -1,16 +1,16 @@
-const path = require('path');
-const getAllFiles = require('../utils/getAllFiles');
+import path from 'path';
+import getAllFiles from '../utils/getAllFiles';
 
-module.exports = (client) => {
+const eventHandler = (client: any) => {
   const eventFolders = getAllFiles(path.join(__dirname, '..', 'events'), true);
 
   for (const eventFolder of eventFolders) {
     const eventFiles = getAllFiles(eventFolder);
-    eventFiles.sort((a, b) => a > b);
+    eventFiles.sort();
 
-    const eventName = eventFolder.replace(/\\/g, '/').split('/').pop();
+    const eventName = eventFolder.replace(/\\/g, '/').split('/').pop()!;
 
-    client.on(eventName, async (arg) => {
+    client.on(eventName, async (arg: any) => {
       for (const eventFile of eventFiles) {
         const eventFunction = require(eventFile);
         await eventFunction(client, arg);
@@ -18,3 +18,5 @@ module.exports = (client) => {
     });
   }
 };
+
+export default eventHandler;
